api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-multipart>=0.0.18
pydantic>=2.10.0
openai>=1.60.0
chromadb>=0.6.0
pypdf>=5.0.0
python-dotenv>=1.0.1
httpx>=0.28.0
